+++ src/Modifiers/BackgroundModifier.php
@@ -4,10 +4,6 @@ declare(strict_types=1);
 
 namespace Intervention\Image\Modifiers;
 
-use Intervention\Image\Colors\Rgb\Channels\Blue;
-use Intervention\Image\Colors\Rgb\Channels\Green;
-use Intervention\Image\Colors\Rgb\Channels\Red;
-use Intervention\Image\Colors\Rgb\Color;
 use Intervention\Image\Drivers\SpecializableModifier;
 use Intervention\Image\Interfaces\ColorInterface;
 use Intervention\Image\Interfaces\DriverInterface;
@@ -25,25 +21,12 @@ class BackgroundModifier extends SpecializableModifier
     }
 
     /**
-     * Decode background color of current modifier with given driver. Possible
-     * (semi-)transparent alpha channel values are made opaque.
+     * Decode background color of current modifier with given driver
      */
     protected function backgroundColor(DriverInterface $driver): ColorInterface
     {
-        // decode background color
-        $color = $driver->handleColorInput(
+        return $driver->handleColorInput(
             $this->color !== null ? $this->color : $driver->config()->backgroundColor
         );
-
-        // replace alpha channel value with opaque value
-        if ($color->isTransparent()) {
-            return new Color(
-                $color->channel(Red::class)->value(),
-                $color->channel(Green::class)->value(),
-                $color->channel(Blue::class)->value(),
-            );
-        }
-
-        return $color;
     }
 }